위젯 클릭 시 실행할 액티비티 추가, AppNotificationManager ui모듈로 이동, 위젯 업데이트 후 RemoteView적용하는 로직 추가 필요

+++ core/database/src/main/java/io/github/pknujsp/weatherwizard/core/database/widget/WidgetDao.kt
@@ -16,12 +16,12 @@ interface WidgetDao {
     fun getAll(): Flow<List<WidgetDto>>
 
     @Query("SELECT * FROM widgets WHERE id = :id")
-    suspend fun getById(id: Long): WidgetDto
+    suspend fun getById(id: Int): WidgetDto
 
     @Query("DELETE FROM widgets WHERE id = :id")
-    suspend fun deleteById(id: Long)
+    suspend fun deleteById(id: Int)
 
     @Query("SELECT EXISTS(SELECT * FROM widgets WHERE id = :id)")
-    suspend fun containsId(id: Long): Boolean
+    suspend fun containsId(id: Int): Boolean
 
 }
